# backend/api/app/schemas/auth.py
"""Auth boundary schemas."""

from pydantic import BaseModel, Field


class LoginIn(BaseModel):
    username: str = Field(min_length=1, max_length=64)
    password: str = Field(min_length=1, max_length=128)


class LoginOut(BaseModel):
    ok: bool = True


class MeOut(BaseModel):
    username: str